evolution tests: rustfmt cleanup + struct-update fix in stagnation test
The auto-formatter rewrote the manual let-mut-assign-fields pattern
in stagnation_termination.rs into Budget struct-update syntax, which
inadvertently shadowed the  binding the assertion was
referencing. Restored the binding so both the assertion and the
panic message can read budget.stagnation_limit.

Also: fmt-only changes to diversity_score.rs (Budget import) and
id_remapping.rs / checkpoint_roundtrip.rs (no semantic change).

cargo test -p wafrift-evolution --test stagnation_termination: 2/2 green.

+++ crates/evolution/tests/id_remapping.rs
@@ -73,7 +73,7 @@ fn map_elites_grid_fills_through_engine_round_trip() {
     // public surface, but stats.evaluations is incremented inside
     // submit_batch only on the success path.
     assert_eq!(
-        engine.stats.evaluations as usize, total_evals,
+        engine.stats.evaluations, total_evals,
         "submit_batch must update stats.evaluations exactly once per result; \
          pre-fix this still passed because the count is engine-side, but \
          the next assertion is the real bug catcher"
